initiating revert of all old readings implementation
